requirements.txt
librosa
pydub
numpy==1.24.3
scipy

openai-whisper

tensorflow==2.13.0
tensorflow-hub

ollama

streamlit

python-dotenv
Pillow

matplotlib
plotly

pytest
black
flake8

setuptools
wheel
